scene folder config file moved out index file changed to add scenes and run window onload
