+++ src/index.js
@@ -13,33 +13,22 @@ But this hasn't happened yet!
 */
 
 import 'phaser';
+import MainScene from './scenes/MainScene'
+import BgScene from './scenes/BgScene';
+import FgScene from './scenes/FgScene';
+import config from './config/config'
 
-import MainScene from './MainScene'
-import BgScene from './BgScene';
-import FgScene from './FgScene';
+class Game extends Phaser.Game {
+  constructor() {
+    super(config);
+    this.scene.add('BgScene', BgScene)
+    this.scene.add('FgScene', FgScene)
+    this.scene.add('MainScene', MainScene)
+    this.scene.start('MainScene')
 
-
-// This will be our initial configuration for when we create the game instance
-const config = {
-  type: Phaser.AUTO,  // Specify the underlying browser rendering engine (AUTO, CANVAS, WEBGL)
-                      // AUTO will attempt to use WEBGL, but if not available it'll default to CANVAS
-  width: 800,   // Game width in pixels
-  height: 600,  // Game height in pixels
-  physics: {    // Optional: specify physics engine and configuration
-    default: 'arcade',  // A simple and performant physics engine
-    arcade: {
-      gravity: { y: 1500 },  // Game objects will by default be affected by gravity
-      debug: false,
-    },
-  },
-  // Specify the scene(s) up front. Can also dynamically add scenes if desired.
-  // Only the first scene in the array will be automatically run
-  scene: [MainScene, BgScene, FgScene],
-  // This option is to turn off the default behavior of images being automatically sharpened.
-  // Since we're using pixel art, we want every beautiful pixel untouched!
-  render: {
-    pixelArt: true,
   }
-};
+}
 
-const game = new Phaser.Game(config);
+window.onload = function () {
+  window.game = new Game();
+}
